+++ src/pages/Home.tsx
@@ -1,89 +1,59 @@
-import React, { FormEvent, useEffect, useRef, useState } from 'react';
+import React from 'react';
+import { Link } from 'react-router-dom';
 
-import CardList from '../components/CardList';
-import { CardProps } from '../components/Card';
-import { RegistrationsQueryResponse } from '../services/registrations-query/query';
-import SearchBar from '../components/SearchBar';
-import { useRegistrationsQuery } from '../services/registrations-query/query-hook';
-import AwesomeDebouncePromise from 'awesome-debounce-promise';
-
-const Home: React.FC = () => {
-  const query = useRegistrationsQuery();
-
-  const [registrationsResponses, setRegistrationsResponses] = useState<
-    RegistrationsQueryResponse[]
-  >([]);
-  const registrations = registrationsResponses
-    ?.map((res) => res.data)
-    .flat()
-    .map(
-      (reg) =>
-        ({
-          name: reg.name,
-          email: reg.contact.email,
-          phoneNumber: reg.contact.phone && reg.contact.phone[0],
-          status: 'OK',
-          statusColor: 'blue',
-        } as CardProps)
-    );
-
-  const [searchTerm, setSearchTerm] = useState<string>('');
-  const searchBarRef = useRef<HTMLInputElement>(null);
-  const onSubmitSearchTerm = (evt: FormEvent) => {
-    evt.preventDefault();
-    setSearchTerm(searchBarRef.current?.value + '');
-  };
-
-  const pageSize = 3;
-  const [pageParams, setPageParams] = useState({
-    page: 1,
-    pageSize,
-  });
-
-  useEffect(() => {
-    setRegistrationsResponses([]);
-    setPageParams({ page: 1, pageSize });
-  }, [searchTerm]);
-
-  const doQuery = AwesomeDebouncePromise(() => {
-    const { page, pageSize } = pageParams;
-    return query({
-      q: searchTerm + '',
-      skip: (page - 1) * pageSize,
-      take: pageSize,
-    }).then((res) =>
-      setRegistrationsResponses([...registrationsResponses, res])
-    );
-  }, 1000);
-
-  useEffect(() => {
-    doQuery();
-  }, [pageParams]);
-
-  const fetchMore = () => {
-    setPageParams({ ...pageParams, page: pageParams.page + 1 });
-  };
-
+const Home = () => {
   return (
-    <main className="bg-background-default min-h-screen">
-      <div className="p-4 container mx-auto max-w-screen-md">
-        <form className="mb-4" onSubmit={onSubmitSearchTerm}>
-          <SearchBar
-            onChange={(event) => setSearchTerm(event.target.value)}
-            value={searchTerm}
-          />
-        </form>
-        <CardList registrations={registrations} />
-        <div className="flex items-center justify-center mt-4">
-          <button
-            className="px-4 py-2 bg-blue-300 rounded-md text-white"
-            onClick={fetchMore}
-          >
-            Carregar mais
-          </button>
+    <>
+      <div
+        className="h-screen w-screen bg-cover bg-center bg-opacity-80"
+        style={{
+          backgroundImage:
+            'url(https://www.ip.usp.br/site/wp-content/uploads/2016/10/FOTO-NESSA-Biblioteca_site2-Copia.jpg)',
+        }}
+      >
+        <div className="container flex items-center mx-auto px-6 md:px-12 relative z-10 py-24 xl:py-40">
+          <div className="bg-white p-5 md:p-10 rounded-xl border flex flex-col items-center lg:w-3/5 xl:w-2/5 relative space-y-6">
+            <h1 className="text-4xl text-red-400 mt-4 font-bold">CEIP-USP</h1>
+
+            <div className="font-source-sans-pro max-w-max space-y-6">
+              <p className="text-md md:text-lg text-center">
+                Se você já tem agendamento para algum serviço no CEIP, crie aqui
+                seu cadastro para acesso ao prédio.
+              </p>
+              <p className="text-md md:text-lg text-center">
+                <span className="font-bold">IMPORTANTE:</span> Este cadastro não
+                garante atendimento. Para informações sobre os serviços
+                prestados no CEIP e forma de inscrição &nbsp;
+                <a
+                  className="text-blue-900 font-bold"
+                  href="https://www.ip.usp.br/site/servicos-a-comunidade/"
+                  target="_blank"
+                  rel="noopener noreferrer"
+                >
+                  acesse aqui
+                </a>
+              </p>
+            </div>
+
+            <div className="flex flex-col space-y-2">
+              <Link
+                to="#"
+                className="bg-transparent hover:bg-blue-500 text-blue-700 font-bold hover:text-white py-2 px-4 border border-blue-500 hover:border-transparent rounded duration-300 text-center"
+                onClick={() => alert('Em Construção!')}
+              >
+                Login
+              </Link>
+              <Link
+                to="/pre-registration"
+                className="bg-transparent hover:bg-blue-500 text-blue-700 font-bold hover:text-white py-2 px-4 border border-blue-500 hover:border-transparent rounded duration-300 text-center"
+              >
+                Realize o seu cadastro
+              </Link>
+            </div>
+          </div>
         </div>
       </div>
-    </main>
+    </>
   );
 };
