+++ src/routes.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import PreRegistration from './pages/PreRegistration/PreRegistration';
+import Search from './pages/Search';
 import Home from './pages/Home';
 
 export default function Routes() {
@@ -8,6 +9,7 @@ export default function Routes() {
     <BrowserRouter>
       <Switch>
         <Route component={Home} path="/" exact />
+        <Route component={Search} path="/search" />
         <Route component={PreRegistration} path="/pre-registration" />
       </Switch>
     </BrowserRouter>
